Proxy cache URL fails match when part UrlEncoded
If I use browsermob-proxy to open url `http://example.com?href=http%3A%2F%2Fexample.com`
Then the URL that appears in the cache is actually `http://example.com?href=http://example.com` which is technically wrong as it contains characters that should be encoded. But as we can't fix that here, we're stuck trying to bodge the situation.